syscallcompat: move getSupplementaryGroups to suser_linux.go
It's only used there, so move it.
